db: move CheckConsistency
Moving CheckConsistency from the manifest package to the main package.
The manifest code shouldn't interact with files/objects on disk.
